+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### ✨ Added
+- handle nested for-loop
+
+### 🔥 Removed
+- possibility to define multiple states for the same route
 
 ## [2.1.5] - 2021-06-22
 ### 🐛 Fixed